+++ components/modules/firebase/SuperAdminCheck.js
@@ -3,7 +3,7 @@ import { fetchCurrUserId } from './fetchCurrUserId';
 
 export const SuperAdminCheck =async () => {
 
-        const currentId = await fetchCurrUserId();
-        console.log(currentId)
-        return(currentId === 'QI0vWII3CkGdS3v8GrKY')
+        const data = await fetchCurrUserId();
+        console.log(data.userId)
+        return(data.userId === 'QI0vWII3CkGdS3v8GrKY')
 }
